fix(senpi-task): restrict Gemini 3.6 provider fallback
